requirements.txt
numpy
stl-numpy